init new-layout
